整理main包，把所有初始化任务抽离出来。尝试集成vercel的api请求模式。
