+++ api/router.go
@@ -1,0 +1,17 @@
+package api
+
+import (
+	"aurora/initialize"
+	"github.com/gin-gonic/gin"
+	"net/http"
+)
+
+var router *gin.Engine
+
+func init() {
+	router = initialize.RegisterRouter()
+}
+
+func Listen(w http.ResponseWriter, r *http.Request) {
+	router.ServeHTTP(w, r)
+}
