+++ main.go
@@ -1,10 +1,7 @@
 package main
 
 import (
-	"aurora/internal/proxys"
-	"bufio"
-	"log/slog"
-	"net/url"
+	"aurora/initialize"
 	"os"
 
 	"github.com/acheong08/endless"
@@ -12,79 +9,9 @@ import (
 	"github.com/joho/godotenv"
 )
 
-func checkProxy() *proxys.IProxy {
-	var proxies []string
-	proxyUrl := os.Getenv("PROXY_URL")
-	if proxyUrl != "" {
-		proxies = append(proxies, proxyUrl)
-	}
-
-	if _, err := os.Stat("proxies.txt"); err == nil {
-		file, _ := os.Open("proxies.txt")
-		defer file.Close()
-		scanner := bufio.NewScanner(file)
-		for scanner.Scan() {
-			proxy := scanner.Text()
-			parsedURL, err := url.Parse(proxy)
-			if err != nil {
-				slog.Warn("proxy url is invalid", "url", proxy, "err", err)
-				continue
-			}
-
-			// 如果缺少端口信息，不是完整的代理链接
-			if parsedURL.Port() != "" {
-				proxies = append(proxies, proxy)
-			} else {
-				continue
-			}
-		}
-	}
-
-	if len(proxies) == 0 {
-		proxy := os.Getenv("http_proxy")
-		if proxy != "" {
-			proxies = append(proxies, proxy)
-		}
-	}
-
-	proxyIP := proxys.NewIProxyIP(proxies)
-	return &proxyIP
-}
-
-func registerRouter() *gin.Engine {
-	handler := NewHandle(
-		checkProxy(),
-		readAccessToken(),
-	)
-
-	router := gin.Default()
-	router.Use(cors)
-
-	router.GET("/", func(c *gin.Context) {
-		c.JSON(200, gin.H{
-			"message": "Hello, world!",
-		})
-	})
-
-	router.GET("/ping", func(c *gin.Context) {
-		c.JSON(200, gin.H{
-			"message": "pong",
-		})
-	})
-
-	router.POST("/auth/session", handler.session)
-	router.POST("/auth/refresh", handler.refresh)
-	router.OPTIONS("/v1/chat/completions", optionsHandler)
-
-	authGroup := router.Group("").Use(Authorization)
-	authGroup.POST("/v1/chat/completions", handler.nightmare)
-	authGroup.GET("/v1/models", handler.engines)
-	return router
-}
-
 func main() {
 	gin.SetMode(gin.ReleaseMode)
-	router := registerRouter()
+	router := initialize.RegisterRouter()
 
 	_ = godotenv.Load(".env")
 	host := os.Getenv("SERVER_HOST")
