+++ middlewares/auth.go
@@ -1,18 +1,10 @@
-package main
+package middlewares
 
 import (
+	"github.com/gin-gonic/gin"
 	"os"
 	"strings"
-
-	gin "github.com/gin-gonic/gin"
 )
-
-func cors(c *gin.Context) {
-	c.Header("Access-Control-Allow-Origin", "*")
-	c.Header("Access-Control-Allow-Methods", "*")
-	c.Header("Access-Control-Allow-Headers", "*")
-	c.Next()
-}
 
 func Authorization(c *gin.Context) {
 	customer_key := os.Getenv("Authorization")
