+++ middlewares/cors.go
@@ -1,0 +1,10 @@
+package middlewares
+
+import "github.com/gin-gonic/gin"
+
+func Cors(c *gin.Context) {
+	c.Header("Access-Control-Allow-Origin", "*")
+	c.Header("Access-Control-Allow-Methods", "*")
+	c.Header("Access-Control-Allow-Headers", "*")
+	c.Next()
+}
